rename AutoApp to TigrblApp and update imports
